Don't rebuild the toolpath on each optimisation iteration
